Consolidate example: simplify_udaf_expression.rs into advanced_udaf.rs
